+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `confidential` client type, to comply with RFC6749 terminology, `private` client type is deprecated
 - Typescript type definitions
 
+### Changed
+- `users.read` scope is added only with `skipUserProfile` option disabled 
 
 ## 1.1.0 - 2022-10-21
 ### Added